feat: Add ErrorBoundary component for handling route errors
