+++ apps/frontend/app/components/dashboard/error-boundary.tsx
@@ -1,0 +1,90 @@
+import {
+	Anchor,
+	Button,
+	Code,
+	Container,
+	Group,
+	List,
+	Stack,
+	Text,
+} from "@mantine/core";
+import { Form, isRouteErrorResponse, useRouteError } from "react-router";
+import { $path } from "safe-routes";
+import { discordLink } from "~/components/dashboard/utils";
+
+export function ErrorBoundary() {
+	const error = useRouteError() as Error;
+	const message = isRouteErrorResponse(error)
+		? error.data.message
+		: error.message;
+
+	return (
+		<Container size="sm" py={{ base: 100, md: 200 }}>
+			<Stack p={{ base: "sm", md: "xl" }}>
+				<Text c="red" fz={{ base: 30, md: 40 }}>
+					We encountered an error
+				</Text>
+				{message ? (
+					<Code mah={100} c="pink">
+						{message}
+					</Code>
+				) : null}
+				<Group wrap="nowrap">
+					<Button
+						fullWidth
+						color="green"
+						variant="outline"
+						onClick={() => window.location.reload()}
+					>
+						Reload
+					</Button>
+					<Form
+						replace
+						method="POST"
+						style={{ width: "100%" }}
+						action={$path("/actions", { intent: "logout" })}
+					>
+						<Button type="submit" variant="outline" color="blue" fullWidth>
+							Logout
+						</Button>
+					</Form>
+				</Group>
+				{isRouteErrorResponse(error) ? null : (
+					<>
+						<Text>This could be due to several reasons:</Text>
+						<List>
+							<List.Item>Your login session has expired/revoked.</List.Item>
+							<List.Item>
+								You don't have permission to perform this action.
+							</List.Item>
+							<List.Item>There was a backend server error.</List.Item>
+						</List>
+						<Text>
+							In most cases, logging out and then logging back in should fix the
+							issue.
+						</Text>
+						<Text>
+							If the error still persists please contact the developer on{" "}
+							<Anchor
+								target="_blank"
+								href={discordLink}
+								rel="noreferrer noopener"
+							>
+								Discord
+							</Anchor>{" "}
+							or create an issue on{" "}
+							<Anchor
+								target="_blank"
+								rel="noreferrer noopener"
+								href="https://github.com/ignisda/ryot/issues"
+							>
+								Github
+							</Anchor>
+							.
+						</Text>
+					</>
+				)}
+			</Stack>
+		</Container>
+	);
+}
